refactor(arch): relocate pure non-UI logic out of UI.Abstractions into TianWen.Lib
Continues the StretchSolver dedup sweep. Two pure-logic units were filed under
UI.Abstractions by accident; move them to where the layering says they belong:

- TransformFactory: Profile -> SOFA.Transform factory, shared by CLI
  (LocationResolver) AND GUI, zero UI deps. -> TianWen.Lib.Devices. All callers
  already imported TianWen.Lib.Devices; dropped the now-dead UI.Abstractions
  using from LocationResolver (IConsoleHost resolves via parent TianWen.Cli ns).

- AutoWhiteBalance gray-world math -> TianWen.Lib.Imaging.GrayWorldWhiteBalance,
  next to BackgroundNeutralization/StretchSolver (the StretchSolver/AstroImageDocument
  pattern). Min/MaxMultiplier + LitFraction + GrayWorldRgb/GrayWorldBayer/Normalize
  are pure span math; the constants had to move with Normalize's clamp since Lib
  can't reference UI. AutoWhiteBalance is now just the IPreviewSource adapter that
  dispatches into Lib. WB-slider range in ImageRendererBase.InfoPanel now reads the
  Lib constant (single source of truth preserved). Test renamed
  AutoWhiteBalanceTests -> GrayWorldWhiteBalanceTests (it only tested the pure
  methods, now a Lib type).

SkyMapProjection deliberately left in UI.Abstractions: its only non-UI consumer is
a test (already runs there, Lib.Tests refs UI.Abstractions), no CLI/headless caller
needs the display-stereographic projection, and moving it would split StarRadius/
StarColor + drag SkyMapState.RaDecToUnitVec for no gain.

Build: CLI + GUI + Lib.Tests 0 warnings/0 errors. 31 affected tests pass.
